# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Akiba - Chama management platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChamaMeetingFrequency = typeof ChamaMeetingFrequency[keyof typeof ChamaMeetingFrequency];


export const ChamaMeetingFrequency = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
} as const;

export interface Chama {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  meetingFrequency: ChamaMeetingFrequency;
  contributionAmount: number;
  /** @nullable */
  bankAccount?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type ChamaInputMeetingFrequency = typeof ChamaInputMeetingFrequency[keyof typeof ChamaInputMeetingFrequency];


export const ChamaInputMeetingFrequency = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
} as const;

export interface ChamaInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  meetingFrequency: ChamaInputMeetingFrequency;
  /** @minimum 1 */
  contributionAmount: number;
  bankAccount?: string;
}

export type ChamaUpdateMeetingFrequency = typeof ChamaUpdateMeetingFrequency[keyof typeof ChamaUpdateMeetingFrequency];


export const ChamaUpdateMeetingFrequency = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
} as const;

export interface ChamaUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  meetingFrequency?: ChamaUpdateMeetingFrequency;
  /** @minimum 1 */
  contributionAmount?: number;
  bankAccount?: string;
}

export type ContributionStatus = typeof ContributionStatus[keyof typeof ContributionStatus];


export const ContributionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Contribution {
  id: number;
  chamaId: number;
  memberId: number;
  memberName: string;
  amount: number;
  status: ContributionStatus;
  /** @nullable */
  mpesaRef?: string | null;
  /** @nullable */
  paidAt: string | null;
  createdAt: string;
}

export interface ChamaSummary {
  chamaId: number;
  totalSavings: number;
  totalLoaned: number;
  totalRepaid: number;
  activeLoanCount: number;
  memberCount: number;
  recentContributions: Contribution[];
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  chairperson: 'chairperson',
  treasurer: 'treasurer',
  secretary: 'secretary',
  member: 'member',
} as const;

export interface Member {
  id: number;
  chamaId: number;
  name: string;
  phoneNumber: string;
  /** @nullable */
  email?: string | null;
  role: MemberRole;
  joinedAt: string;
  createdAt?: string;
}

export type MemberInputRole = typeof MemberInputRole[keyof typeof MemberInputRole];


export const MemberInputRole = {
  chairperson: 'chairperson',
  treasurer: 'treasurer',
  secretary: 'secretary',
  member: 'member',
} as const;

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phoneNumber: string;
  email?: string;
  role: MemberInputRole;
}

export type MemberUpdateRole = typeof MemberUpdateRole[keyof typeof MemberUpdateRole];


export const MemberUpdateRole = {
  chairperson: 'chairperson',
  treasurer: 'treasurer',
  secretary: 'secretary',
  member: 'member',
} as const;

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  phoneNumber?: string;
  email?: string;
  role?: MemberUpdateRole;
}

export type ContributionInputStatus = typeof ContributionInputStatus[keyof typeof ContributionInputStatus];


export const ContributionInputStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ContributionInput {
  memberId: number;
  /** @minimum 1 */
  amount: number;
  mpesaRef?: string;
  status?: ContributionInputStatus;
}

export type LoanStatus = typeof LoanStatus[keyof typeof LoanStatus];


export const LoanStatus = {
  pending: 'pending',
  active: 'active',
  repaid: 'repaid',
  defaulted: 'defaulted',
} as const;

export interface Loan {
  id: number;
  chamaId: number;
  memberId: number;
  memberName: string;
  principal: number;
  interestRate: number;
  totalDue?: number;
  totalRepaid?: number;
  status: LoanStatus;
  /** @nullable */
  disbursedAt: string | null;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface LoanInput {
  memberId: number;
  /** @minimum 1 */
  principal: number;
  /** @minimum 0 */
  interestRate: number;
  dueDate: string;
  notes?: string;
}

export type LoanUpdateStatus = typeof LoanUpdateStatus[keyof typeof LoanUpdateStatus];


export const LoanUpdateStatus = {
  pending: 'pending',
  active: 'active',
  repaid: 'repaid',
  defaulted: 'defaulted',
} as const;

export interface LoanUpdate {
  status?: LoanUpdateStatus;
  disbursedAt?: string;
  notes?: string;
}

export interface Repayment {
  id: number;
  loanId: number;
  chamaId: number;
  memberId: number;
  amount: number;
  /** @nullable */
  mpesaRef?: string | null;
  paidAt: string;
  createdAt: string;
}

export interface RepaymentInput {
  /** @minimum 1 */
  amount: number;
  mpesaRef?: string;
}

export interface MpesaPaymentInput {
  /** @minLength 1 */
  phoneNumber: string;
  /** @minimum 1 */
  amount: number;
  /** @minLength 1 */
  accountReference: string;
  /** @minLength 1 */
  transactionDesc: string;
}

export interface MpesaPaymentResult {
  merchantRequestId: string;
  checkoutRequestId: string;
  responseCode: string;
  responseDescription: string;
  customerMessage: string;
}

export type MpesaCallbackPayloadBody = { [key: string]: unknown };

export interface MpesaCallbackPayload {
  Body?: MpesaCallbackPayloadBody;
}