# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Akiba - Chama management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chamas:
    get:
      operationId: listChamas
      tags: [chamas]
      summary: List all chamas
      responses:
        "200":
          description: List of chamas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chama"
    post:
      operationId: createChama
      tags: [chamas]
      summary: Create a new chama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChamaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chama"
  /chamas/{id}:
    get:
      operationId: getChama
      tags: [chamas]
      summary: Get a chama by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chama details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chama"
        "404":
          description: Not found
    patch:
      operationId: updateChama
      tags: [chamas]
      summary: Update a chama
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChamaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chama"
        "404":
          description: Not found
    delete:
      operationId: deleteChama
      tags: [chamas]
      summary: Delete a chama
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /chamas/{id}/summary:
    get:
      operationId: getChamaSummary
      tags: [chamas]
      summary: Get financial summary for a chama
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chama financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChamaSummary"
        "404":
          description: Not found
...
  /chamas/{chamaId}/contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contributions for a chama
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of contributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
    post:
      operationId: recordContribution
      tags: [contributions]
      summary: Record a contribution manually
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
  /chamas/{chamaId}/loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans for a chama
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a loan for a member
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /chamas/{chamaId}/loans/{loanId}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get a loan
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update loan status or details
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
  /chamas/{chamaId}/loans/{loanId}/repayments:
    get:
      operationId: listRepayments
      tags: [loans]
      summary: List repayments for a loan
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of repayments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repayment"
    post:
      operationId: recordRepayment
      tags: [loans]
      summary: Record a loan repayment
      parameters:
        - name: chamaId
          in: path
          required: true
          schema:
            type: integer
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repayment"
...
components:
  schemas:
...
    Chama:
      type: object
      required: [id, name, meetingFrequency, contributionAmount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        meetingFrequency:
          type: string
          enum: [weekly, monthly, quarterly]
        contributionAmount:
          type: number
        bankAccount:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChamaInput:
      type: object
      required: [name, meetingFrequency, contributionAmount]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        meetingFrequency:
          type: string
          enum: [weekly, monthly, quarterly]
        contributionAmount:
          type: number
          minimum: 1
        bankAccount:
          type: string
    ChamaUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        meetingFrequency:
          type: string
          enum: [weekly, monthly, quarterly]
        contributionAmount:
          type: number
          minimum: 1
        bankAccount:
          type: string
    ChamaSummary:
      type: object
      required: [chamaId, totalSavings, totalLoaned, totalRepaid, activeLoanCount, memberCount, recentContributions]
      properties:
        chamaId:
          type: integer
        totalSavings:
          type: number
        totalLoaned:
          type: number
        totalRepaid:
          type: number
        activeLoanCount:
          type: integer
        memberCount:
          type: integer
        recentContributions:
          type: array
          items:
            $ref: "#/components/schemas/Contribution"
...
    Contribution:
      type: object
      required: [id, chamaId, memberId, memberName, amount, status, paidAt, createdAt]
      properties:
        id:
          type: integer
        chamaId:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        mpesaRef:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ContributionInput:
      type: object
      required: [memberId, amount]
      properties:
        memberId:
          type: integer
        amount:
          type: number
          minimum: 1
        mpesaRef:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
    Loan:
      type: object
      required: [id, chamaId, memberId, memberName, principal, interestRate, status, disbursedAt, dueDate, createdAt]
      properties:
        id:
          type: integer
        chamaId:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        principal:
          type: number
        interestRate:
          type: number
        totalDue:
          type: number
        totalRepaid:
          type: number
        status:
          type: string
          enum: [pending, active, repaid, defaulted]
        disbursedAt:
          type: ["string", "null"]
          format: date-time
        dueDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    LoanInput:
      type: object
      required: [memberId, principal, interestRate, dueDate]
      properties:
        memberId:
          type: integer
        principal:
          type: number
          minimum: 1
        interestRate:
          type: number
          minimum: 0
        dueDate:
          type: string
          format: date
        notes:
          type: string
    LoanUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, repaid, defaulted]
        disbursedAt:
          type: string
          format: date-time
        notes:
          type: string
    Repayment:
      type: object
      required: [id, loanId, chamaId, memberId, amount, paidAt, createdAt]
      properties:
        id:
          type: integer
        loanId:
          type: integer
        chamaId:
          type: integer
        memberId:
          type: integer
        amount:
          type: number
        mpesaRef:
          type: ["string", "null"]
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    RepaymentInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 1
        mpesaRef:
          type: string
...