# lib/api-zod/src/generated/types/chama.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Akiba - Chama management platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChamaMeetingFrequency } from './chamaMeetingFrequency';

export interface Chama {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  meetingFrequency: ChamaMeetingFrequency;
  contributionAmount: number;
  /** @nullable */
  bankAccount?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}