# lib/api-zod/src/generated/types/chamaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Akiba - Chama management platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChamaInputMeetingFrequency } from './chamaInputMeetingFrequency';

export interface ChamaInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  meetingFrequency: ChamaInputMeetingFrequency;
  /** @minimum 1 */
  contributionAmount: number;
  bankAccount?: string;
}