# lib/api-zod/src/generated/types/contribution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Akiba - Chama management platform
 * OpenAPI spec version: 0.1.0
 */
import type { ContributionStatus } from './contributionStatus';

export interface Contribution {
  id: number;
  chamaId: number;
  memberId: number;
  memberName: string;
  amount: number;
  status: ContributionStatus;
  /** @nullable */
  mpesaRef?: string | null;
  /** @nullable */
  paidAt: Date | null;
  createdAt: Date;
}