# lib/api-zod/src/generated/types/loanUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Akiba - Chama management platform
 * OpenAPI spec version: 0.1.0
 */
import type { LoanUpdateStatus } from './loanUpdateStatus';

export interface LoanUpdate {
  status?: LoanUpdateStatus;
  disbursedAt?: Date;
  notes?: string;
}